// frontend/src/api/generated/models/supplierInvoiceResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface SupplierInvoiceResponse {
  id: string;
  invoice_number: string;
  date: string;
  due_date: string | null;
  total_amount: string;
  amount_paid: string;
  status: string;
}